requirements.txt
flask
torch==2.0.1+cpu
torchvision==0.15.2+cpu
git+https://github.com/openai/CLIP.git
Pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
